requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asttokens==3.0.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
cramjam==2.11.0
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fastparquet==2024.11.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
pure_eval==0.2.3
pyg_lib==0.6.0.dev20251107+pt29cpu
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
torch-geometric==2.7.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
